putOrder теперь умеет обновлять заявку

+++ src/main/java/ru/pashkovske/buratino/tinkoff/service/model/order/OrderCommand.java
@@ -5,6 +5,6 @@ import ru.tinkoff.piapi.contract.v1.OrderDirection;
 
 public interface OrderCommand<T> {
     InstrumentHolder<T> getInstrument();
-    int getLotQuantity();
+    long getLotQuantity();
     OrderDirection getDirection();
 }
